feat: Created whole basic insert flow.

+++ FinSync.Finances/Src/FinSync.Finances.Application/DTOs/FinancialAccountForCreationDTO.cs
@@ -3,5 +3,5 @@ namespace FinSync.Finances.Application.DTOs;
 public record FinancialAccountForCreationDTO(
   string UserId,
   string Name,
-  double Balance,
+  decimal Balance,
   string CurrencyAcronym);
